rewrite engine

+++ src/ext/vk_descriptors.cpp
@@ -1,5 +1,4 @@
 ﻿#include "vk_descriptors.h"
-#include "vk_initializers.h"
 
 #include <iostream>
 #include <exception>
@@ -182,7 +181,7 @@ void DescriptorAllocatorGrowable::init(VkDevice device, uint32_t maxSets, std::s
 
     VkDescriptorPool newPool = create_pool(device, maxSets, poolRatios);
 
-    setsPerPool = maxSets * 1.5; //grow it next allocation
+    setsPerPool = static_cast<uint32_t>(maxSets * 1.5); //grow it next allocation
 
     readyPools.push_back(newPool);
 }
@@ -231,7 +230,7 @@ VkDescriptorPool DescriptorAllocatorGrowable::get_pool(VkDevice device)
         //need to create a new pool
         newPool = create_pool(device, setsPerPool, ratios);
 
-        setsPerPool = setsPerPool * 1.5;
+        setsPerPool = static_cast<uint32_t>(setsPerPool * 1.5);
         if (setsPerPool > 4092)
         {
             setsPerPool = 4092;
